requirements.txt
python-binance>=1.0.28
pandas>=1.3.0
pandas-ta>=0.3.14b
numpy==1.21.6
matplotlib>=3.4.0
backtesting>=0.3.3
pyyaml>=6.0
python-dotenv>=0.19.0
requests>=2.25.0
ccxt>=2.0.0
